deps/v8/test/inspector/runtime/console-time-end-format-expected.txt

Checks format of console.timeEnd output

Running test: zero
js: 0 ms
timeEnd: 0 ms

Running test: verySmall
js: 1e-15 ms
timeEnd: 1e-15 ms

Running test: small
js: 0.001 ms
timeEnd: 0.001 ms

Running test: regular
js: 1.2345 ms
timeEnd: 1.2345 ms

Running test: big
js: 10000.2345 ms
timeEnd: 10000.2345 ms

Running test: veryBig
js: 1000000000000000.2 ms
timeEnd: 1000000000000000.2 ms

Running test: huge
js: 1e+42 ms
timeEnd: 1e+42 ms

Running test: undefinedAsLabel
js: 1 ms
default: 1 ms

Running test: emptyAsLabel
js: 1 ms
default: 1 ms
